% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measles-concentration.R
\name{calc_concentration}
\alias{calc_concentration}
\title{Helper function to calculate concentration}
\usage{
calc_concentration(year_in)
}
\arguments{
\item{year_in}{the year in which concentration is calculated}
}
\value{
a wrangled data set
}
\description{
Helper function to calculate concentration
}
