% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdp_ttest.R
\name{gdp_ttest}
\alias{gdp_ttest}
\title{Creates a table showing the results of a two sample t-test on measles and GDP growth}
\usage{
gdp_ttest(year_in)
}
\arguments{
\item{year_in}{a single year}
}
\value{
a table with 10 columns and one row of test results
}
\description{
Creates a table showing the results of a two sample t-test on measles and GDP growth
}
