% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data_month}
\alias{load_data_month}
\title{Loads in original monthly measles data set}
\usage{
load_data_month()
}
\value{
monthly measles data sets
}
\description{
Loads in original monthly measles data set
}
