% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data_year}
\alias{load_data_year}
\title{Loads in original yearly measles data set}
\usage{
load_data_year()
}
\value{
yearly measles data set
}
\description{
Loads in original yearly measles data set
}
