% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measles_covid_barplot.R
\name{measles_covid_barplot}
\alias{measles_covid_barplot}
\title{Creates a horizontal bar plot comparing measles and Covid-19 cases by country}
\usage{
measles_covid_barplot(year)
}
\arguments{
\item{year}{a single year}
}
\value{
a horizontal bar plot with two bars for each country
}
\description{
Creates a horizontal bar plot comparing measles and Covid-19 cases by country
}
