% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measles_vs_covid.R
\name{measles_covid_compare}
\alias{measles_covid_compare}
\title{Helper function to compare measles and Covid-19 cases in a select year}
\usage{
measles_covid_compare(year1)
}
\arguments{
\item{year1}{a single year}
}
\value{
a wrangled data set
}
\description{
Helper function to compare measles and Covid-19 cases in a select year
}
