% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measles_vs_covid.R
\name{measles_vs_covid}
\alias{measles_vs_covid}
\title{Creates a table comparing measles and Covid-19 cases in two years}
\usage{
measles_vs_covid(year1, year2)
}
\arguments{
\item{year1}{first comparison year}

\item{year2}{second comparison year}
}
\value{
a formatted table comparing number of cases for two years
}
\description{
Creates a table comparing measles and Covid-19 cases in two years
}
