% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measles-concentration.R
\name{summarize_concentration}
\alias{summarize_concentration}
\title{Helper function to summarize concentration}
\usage{
summarize_concentration(yr)
}
\arguments{
\item{yr}{a single year}
}
\value{
a wrangled data set
}
\description{
Helper function to summarize concentration
}
