% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measles-concentration.R
\name{measles_concentration}
\alias{measles_concentration}
\title{Creates a table a table showing the concentration of measles cases}
\usage{
measles_concentration(start_year, end_year)
}
\arguments{
\item{start_year}{starting year for analysis}

\item{end_year}{ending year for analysis}
}
\value{
a formatted table with 6 columns of information
}
\description{
Creates a table a table showing the concentration of measles cases
}
