% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-measles-heatmap.R
\name{plot_measles_heatmap}
\alias{plot_measles_heatmap}
\title{Creates a heatmap plot of countries with the highest concentration of cases}
\usage{
plot_measles_heatmap(start_year, end_year)
}
\arguments{
\item{start_year}{starting year for analysis}

\item{end_year}{ending year for analysis}
}
\value{
a styled heatmap including 15 countries
}
\description{
Creates a heatmap plot of countries with the highest concentration of cases
}
